requirements.txt
Django>=3.2
djangorestframework>=3.12
openai>=0.8.0
